requirements.txt
asgiref==3.5.1
attrs==21.4.0
autobahn==22.5.1
Automat==20.2.0
certifi==2022.5.18
cffi==1.15.0
channels==3.0.4
cloudinary==1.29.0
constantly==15.1.0
cryptography==37.0.2
daphne==3.0.2
Django==4.0.4
django-ckeditor==6.4.2
django-cors-headers==3.12.0
django-filter==21.1
django-js-asset==2.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
dnspython==2.2.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
Markdown==3.3.7
msgpack==1.0.4
Pillow==9.1.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
python-snappy==0.6.1
pytz==2022.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.2.0
urllib3==1.26.9
zope.interface==5.4.0
